refactor:Train repository 수정 pk값 변경

+++ src/main/java/com/example/planergram/repository/TrainRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TrainRepository extends JpaRepository<Train, Integer> {
+public interface TrainRepository extends JpaRepository<Train, Long> {
     @Query("from Train where endPoint= :endPoint")
     List<Train> findBySelected(@Param("endPoint") String endPoint);
 }
